renamed espnow files and did some style changes
